+++ lib/ui/notifications/NotificationsHomePanel.dart
@@ -326,14 +326,14 @@ class _NotificationsHomePanelState extends State<NotificationsHomePanel> impleme
             onTap: _onTapFilter,
             child: Container(
                 decoration: BoxDecoration(
-                    color: Styles().colors.surface,
-                    border: Border.all(color: Styles().colors.disabledTextColor, width: 1),
+                    color: Styles().colors.background,
+                    border: Border.all(color: Styles().colors.fillColorSecondary, width: 1),
                     borderRadius: BorderRadius.circular(18)),
                 child: Padding(
                     padding: EdgeInsets.symmetric(vertical: 3, horizontal: 8),
                     child: Wrap(crossAxisAlignment: WrapCrossAlignment.center, children: [
                       Padding(padding: EdgeInsets.only(right: 6), child: Styles().images.getImage('filters')),
-                      Text(title, style: Styles().textStyles.getTextStyle('widget.button.title.regular'), semanticsLabel: ''),
+                      Text(title, style: Styles().textStyles.getTextStyle('widget.button.light.title.regular'), semanticsLabel: ''),
                       Padding(padding: EdgeInsets.only(left: 3), child: Styles().images.getImage('chevron-right'))
                     ])))));
   }
@@ -523,9 +523,11 @@ class _NotificationsHomePanelState extends State<NotificationsHomePanel> impleme
         padding: EdgeInsets.only(top: 30),
         child: RoundedButton(
             label: Localization().getStringEx('panel.inbox.filter.apply.button', 'Apply'),
+            textStyle: Styles().textStyles.getTextStyle('widget.button.title.medium.fat'),
+            backgroundColor: Styles().colors.background,
+            borderColor: Styles().colors.fillColorSecondary,
             padding: EdgeInsets.symmetric(vertical: 4),
             contentWeight: 0.35,
-            fontSize: 16,
             onTap: _onTapApplyFilter));
   }