+++ CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased - NEOM U
+### Changed
+- Update UI to use dark theme (progress) [#80](https://github.com/rokmetro/neom-app/issues/80)
 
 ## [0.5.14] = 2025-03-19
 ### Changed
-- Update UI to use dark theme [#80](https://github.com/rokmetro/neom-app/issues/80)
+- Update UI to use dark theme (progress) [#80](https://github.com/rokmetro/neom-app/issues/80)
 
 ## [0.5.13] = 2025-03-14
 ### Added